<REFACTOR> LadderFactory 추상화
